Update minutes to years and days.java

+++ minutes to years and days.java
@@ -5,7 +5,7 @@ public class Projecttwo {
 	public static void main(String[] args) {
 		System.out.println("enter minutes");
 		Scanner sc =new Scanner(System.in);
-		float a= sc.nextInt();
+		float a= sc.nextFloat();
 		float b=(float) (a/(60*24*365));
 		int c= (int) b;
 		float d=(b-c)*365;
